refactor: extract credit note form items into a specific component
