+++ src/pages/CreateCreditNote.tsx
@@ -1,7 +1,5 @@
 import { gql } from '@apollo/client'
 import { useFormik } from 'formik'
-import _get from 'lodash/get'
-import { DateTime } from 'luxon'
 import { useEffect, useMemo, useRef, useState } from 'react'
 import { generatePath, useNavigate, useParams } from 'react-router-dom'
 import styled from 'styled-components'
@@ -9,8 +7,8 @@ import { array, object, string } from 'yup'
 
 import { CreditNoteCodeSnippet } from '~/components/creditNote/CreditNoteCodeSnippet'
 import { CreditNoteFormCalculation } from '~/components/creditNote/CreditNoteFormCalculation'
-import { CreditNoteFormItem } from '~/components/creditNote/CreditNoteFormItem'
-import { CreditNoteForm, CreditTypeEnum, FromFee, GroupedFee } from '~/components/creditNote/types'
+import { CreditNoteItemsForm } from '~/components/creditNote/CreditNoteItemsForm'
+import { CreditNoteForm, CreditTypeEnum } from '~/components/creditNote/types'
 import { creditNoteFormCalculationCalculation } from '~/components/creditNote/utils'
 import {
   Alert,
@@ -24,7 +22,7 @@ import {
   StatusType,
   Typography,
 } from '~/components/designSystem'
-import { Checkbox, ComboBoxField, TextInputField } from '~/components/form'
+import { ComboBoxField, TextInputField } from '~/components/form'
 import { WarningDialog, WarningDialogRef } from '~/components/WarningDialog'
 import { hasDefinedGQLError } from '~/core/apolloClient'
 import { CustomerInvoiceDetailsTabsOptionsEnum } from '~/core/constants/NavigationEnum'
@@ -46,7 +44,7 @@ import {
 } from '~/generated/graphql'
 import { useInternationalization } from '~/hooks/core/useInternationalization'
 import { useCreateCreditNote } from '~/hooks/useCreateCreditNote'
-import { HEADER_TABLE_HEIGHT, PageHeader, theme } from '~/styles'
+import { PageHeader, theme } from '~/styles'
 import { Content, Main, Side, Subtitle, Title } from '~/styles/mainObjectsForm'
 
 gql`
@@ -93,20 +91,6 @@ export const CREDIT_NOTE_REASONS: { reason: CreditNoteReasonEnum; label: string 
     label: 'text_636d86201507276b7421a981',
   },
 ]
-
-const determineCheckboxValue = (
-  initialValue: boolean | undefined | null,
-  additionnalValue: boolean | undefined,
-) => {
-  if (initialValue === undefined || additionnalValue === undefined) return undefined
-  if (initialValue === null) {
-    return additionnalValue
-  }
-  if (initialValue !== additionnalValue) {
-    return undefined
-  }
-  return additionnalValue
-}
 
 const mapStatus = (type?: InvoicePaymentStatusTypeEnum | undefined): StatusProps => {
   switch (type) {
@@ -191,55 +175,6 @@ const CreateCreditNote = () => {
   })
 
   const hasError = !!formikProps.errors.fees || !!formikProps.errors.addOnFee
-
-  const checkboxGroupValue = useMemo(() => {
-    const fees = formikProps.values.fees || {}
-
-    return (
-      Object.keys(fees).reduce((acc, subscriptionKey) => {
-        const subscriptionValues = fees[subscriptionKey]
-
-        let subscriptionGroupValues: {
-          value: undefined | boolean | null
-          [key: string]: undefined | boolean | null
-        } = {
-          value: null,
-        }
-
-        Object.keys(subscriptionValues.fees).forEach((childKey) => {
-          const child = subscriptionValues.fees[childKey] as FromFee
-
-          if (typeof child?.checked === 'boolean') {
-            subscriptionGroupValues = {
-              ...subscriptionGroupValues,
-              value: determineCheckboxValue(subscriptionGroupValues.value, child?.checked),
-            }
-          } else {
-            let groupValue: boolean | undefined | null = null
-
-            const grouped = (child as unknown as GroupedFee)?.grouped
-
-            Object.keys(grouped || {}).forEach((groupedKey) => {
-              const feeValues = grouped[groupedKey]
-
-              groupValue = determineCheckboxValue(groupValue, feeValues.checked)
-            })
-
-            subscriptionGroupValues = {
-              ...subscriptionGroupValues,
-              [childKey]: groupValue,
-              value: determineCheckboxValue(
-                subscriptionGroupValues.value,
-                groupValue as unknown as boolean | undefined,
-              ),
-            }
-          }
-        })
-
-        return { ...acc, [subscriptionKey]: subscriptionGroupValues }
-      }, {}) || {}
-    )
-  }, [formikProps.values.fees])
 
   const { feeForEstimate } = useMemo(
     () =>
@@ -412,159 +347,14 @@ const CreateCreditNote = () => {
                     </Typography>
                   </div>
 
-                  {isPrepaidCreditsInvoice && (
-                    <HeaderLine className="!mb-0">
-                      <Checkbox
-                        label={'Items'}
-                        value={formikProps.values.creditFee?.[0]?.checked}
-                        onChange={(_, value) => {
-                          formikProps.setFieldValue(`creditFee.0.checked`, value)
-                        }}
-                      />
-
-                      <Typography variant="bodyHl" color="grey500">
-                        {translate('text_636bedf292786b19d3398ee0')}
-                      </Typography>
-                    </HeaderLine>
-                  )}
-
-                  {feeForCredit &&
-                    feeForCredit.map((fee, i) => (
-                      <CreditNoteFormItem
-                        key={fee?.id}
-                        formikProps={formikProps}
-                        currency={currency}
-                        feeName={translate('text_1729262241097k3cnpci6p5j')}
-                        formikKey={`creditFee.${i}`}
-                        maxValue={fee?.maxAmount}
-                      />
-                    ))}
-
-                  {feeForAddOn &&
-                    feeForAddOn.map((fee, i) => (
-                      <CreditNoteFormItem
-                        key={fee?.id}
-                        formikProps={formikProps}
-                        currency={currency}
-                        feeName={fee?.name}
-                        formikKey={`addOnFee.${i}`}
-                        maxValue={fee?.maxAmount}
-                      />
-                    ))}
-
-                  {feesPerInvoice &&
-                    Object.keys(feesPerInvoice).map((subKey) => {
-                      const subscription = feesPerInvoice[subKey]
-
-                      return (
-                        <div key={subKey}>
-                          <HeaderLine>
-                            <Checkbox
-                              value={_get(checkboxGroupValue, `${subKey}.value`)}
-                              canBeIndeterminate
-                              label={
-                                <Typography variant="bodyHl" color="grey500">
-                                  {subscription?.subscriptionName}
-                                </Typography>
-                              }
-                              onChange={(_, value) => {
-                                const childValues = _get(
-                                  formikProps.values.fees,
-                                  `${subKey}.fees`,
-                                ) as unknown as { [feeGroupId: string]: FromFee | GroupedFee }
-
-                                formikProps.setFieldValue(
-                                  `fees.${subKey}.fees`,
-                                  Object.keys(childValues).reduce((acc, childKey) => {
-                                    const child = childValues[childKey] as FromFee
-
-                                    if (typeof child.checked === 'boolean') {
-                                      acc = { ...acc, [childKey]: { ...child, checked: value } }
-                                    } else {
-                                      const grouped = (child as unknown as GroupedFee)?.grouped
-
-                                      acc = {
-                                        ...acc,
-                                        [childKey]: {
-                                          ...child,
-                                          grouped: Object.keys(grouped || {}).reduce(
-                                            (accGroup, groupKey) => {
-                                              const fee = grouped[groupKey]
-
-                                              return {
-                                                ...accGroup,
-                                                [groupKey]: { ...fee, checked: value },
-                                              }
-                                            },
-                                            {},
-                                          ),
-                                        },
-                                      }
-                                    }
-                                    return acc
-                                  }, {}),
-                                )
-                              }}
-                            />
-                            <Typography variant="bodyHl" color="grey500">
-                              {translate('text_636bedf292786b19d3398ee0')}
-                            </Typography>
-                          </HeaderLine>
-                          {Object.keys(subscription?.fees)?.map((groupFeeKey) => {
-                            const child = subscription?.fees[groupFeeKey] as FromFee
-
-                            if (typeof child?.checked === 'boolean') {
-                              return (
-                                <CreditNoteFormItem
-                                  key={child?.id}
-                                  formikProps={formikProps}
-                                  currency={currency}
-                                  feeName={`${child?.name}${
-                                    child.isTrueUpFee
-                                      ? ` - ${translate('text_64463aaa34904c00a23be4f7')}`
-                                      : ''
-                                  }`}
-                                  formikKey={`fees.${subKey}.fees.${groupFeeKey}`}
-                                  maxValue={child?.maxAmount || 0}
-                                  feeSucceededAt={
-                                    !!child?.succeededAt
-                                      ? DateTime.fromISO(child?.succeededAt).toFormat(
-                                          'LLL. dd, yyyy',
-                                        )
-                                      : undefined
-                                  }
-                                />
-                              )
-                            }
-
-                            const grouped = (child as unknown as GroupedFee)?.grouped
-
-                            return (
-                              <div key={groupFeeKey}>
-                                {Object.keys(grouped).map((groupedFeeKey) => {
-                                  const fee = grouped[groupedFeeKey]
-
-                                  return (
-                                    <CreditNoteFormItem
-                                      key={fee?.id}
-                                      formikProps={formikProps}
-                                      currency={currency}
-                                      feeName={`${child.name} • ${fee?.name}${
-                                        fee.isTrueUpFee
-                                          ? ` - ${translate('text_64463aaa34904c00a23be4f7')}`
-                                          : ''
-                                      }`}
-                                      formikKey={`fees.${subKey}.fees.${groupFeeKey}.grouped.${fee?.id}`}
-                                      maxValue={fee?.maxAmount || 0}
-                                    />
-                                  )
-                                })}
-                              </div>
-                            )
-                          })}
-                        </div>
-                      )
-                    })}
+                  <CreditNoteItemsForm
+                    isPrepaidCreditsInvoice={isPrepaidCreditsInvoice}
+                    formikProps={formikProps}
+                    feeForCredit={feeForCredit}
+                    feeForAddOn={feeForAddOn}
+                    feesPerInvoice={feesPerInvoice}
+                    currency={currency}
+                  />
 
                   {isPrepaidCreditsInvoice ? (
                     <>
@@ -659,14 +449,6 @@ const StyledCard = styled.div<{ $loading?: boolean }>`
   > *:last-child {
     margin-left: auto;
   }
-`
-
-const HeaderLine = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  box-shadow: ${theme.shadows[7]};
-  height: ${HEADER_TABLE_HEIGHT}px;
 `
 
 const ButtonContainer = styled.div`
